Use PSR-4-compatible namespaces
